+++ modules/circe/src/test/scala/io/bartholomews/fsclient/samples/v2/AuthorizationCodeGrantExample.scala
@@ -83,7 +83,7 @@ object AuthorizationCodeGrantExample extends App {
 
   // 5. Use the access token
   // an empty request with client `User-Agent` header
-  baseRequest(client)
+  baseRequest(userAgent)
     .get(uri"https://some-server/authenticated-endpoint")
     .sign // sign with the implicit token provided