+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Changed
 - `AccessTokenSigner` using `currentTimeMillis` instead of `nanoTime`
+- `baseRequest` taking an `userAgent` instead of `client`
 ### Fixed
 
 ## [0.1.1] - 2021-01-30